Initial commit
Spending Predictor uses the SciKit model to create a predicted set of data of spending habits for a year. 
It accepts a spreadsheet of 'Date' and 'Spending' columns, and predicts a year's worth of spending based on the patterns and values in the table, and saves the file as a comma delimited file.